Add files via upload
The goal of this project was to look at patterns and trends in public airlines
data and compare them over a period of time (2000-2023). I also wanted to
to see how Covid-19 and The Great Recession impacted the airlines industry.
The main sources of data were taken from the Bureau of Transportation
Statistics. I had to filter some years out, and also make sure that all cost
variables were adjusted for inflation to match the year of 2023 for a more
accurate reading of what the money spent represents in the most present
year. This I adjusted manually with statistic formulas taken from the
Bureau of Labor Statistics Consumer Price Index. I adjusted these for Fuel
Costs (per year in millions), Fuel Cost per Gallon, and Average Ticket Fares.